fix some minor issue

+++ install_pkg.sh
@@ -1,28 +1,28 @@
 #!/bin/sh
 # 脚本能力：
 #  - 安装必备的工具： vim curl python3 python3-pip nginx jq shfmt netcat util-linux tree avahi-daemon
-#  - 安装服务端 vim 写 shell以及配置 nginx 的必备插件 shfmt nginx-config
 #  - 删除 nofound=return in /etc/nsswich.conf
 #  - 配置 ntp 服务器为 ntp.aliyun.com
-#  - 内网穿透能力
 
 # 可能需要根据情况修改 PROXY 变量
 # PROXY="-x socks5h://127.0.0.1:2023"
 # 初始化 adhole 系统的版本
-VER="1.1.5"
+VER="1.1.6"
 # Packages to install on Debian/Ubuntu systems
 # gh=github-cli
 # added libpacparser1 to check wpad proxy setup
 PKGS="vim psmisc curl python3 python3-pip nginx jq git netcat-openbsd util-linux \
     tree parallel avahi-daemon nsd unbound xz-utils lsof zstd \
     bind9-dnsutils network-manager dos2unix libpacparser1 systemd-timesyncd"
-# Modify to suit your own data
+# Modify to suit your own requirement if not in China.
 TZONE="Asia/Shanghai"
 NTP="ntp.aliyun.com"
 #
 echo "Info: 本脚本需要 sudo 能力，检查当前用户是否具有 sudo 能力"
-sudo -nv 2>/dev/null
-[ $? != 0 ] && echo "Error: 当前用户没有 sudo 能力" && exit 1
+if ! sudo -nv 2>/dev/null; then
+	echo "Error: 当前用户没有 sudo 能力"
+	exit 1
+fi
 
 ADHOLE="/etc/adhole_version"
 [ "$(cat $ADHOLE 2>/dev/null)" != "$VER" ] && echo $VER | sudo tee $ADHOLE
@@ -30,7 +30,7 @@ ADHOLE="/etc/adhole_version"
 update_issue() {
 	# [ -r /etc/debian_version ] && DEB_VER=$(cat /etc/debian_version)
 	if ! grep ^adhole /etc/issue 2>/dev/null; then
-	cat <<EOI | sudo tee -a /etc/issue
+		cat <<EOI | sudo tee -a /etc/issue
 adhole $VER \S \l
 EOI
 	fi
@@ -47,14 +47,16 @@ EOSS
 base_line() {
 	# 基线，使用 network-manager，停用 systemd-resolved
 	# sudo dpkg-query -W $PKGS >/dev/null
-  if ! sudo -E apt -y install $PKGS ;then
-		echo "Error: Install PKGs failed!" 
+	if ! sudo -E apt -y install "$PKGS"; then
+		echo "Error: Install PKGs failed!"
 		exit 1
 	fi
 	# 时钟管理非常重要，否则 DNS 不能工作(需要去和 root DNS 同步)
 	# sudo apt -y purge ntp chrony 2>/dev/null
-	sudo systemctl start systemd-timesyncd
-	[ $? != 0 ] && echo "Error: Failed to start systemd-timesyncd" && exit 1
+	if ! sudo systemctl start systemd-timesyncd; then
+		echo "Error: Failed to start systemd-timesyncd"
+		exit 1
+	fi
 	sudo timedatectl set-timezone "$TZONE"
 	sudo sed -i "s/#NTP=/NTP=$NTP/g" /etc/systemd/timesyncd.conf
 	sudo systemctl restart systemd-timesyncd
@@ -74,7 +76,7 @@ base_line() {
 	else
 		sudo sed -i '/^\[main/a dns=none' /etc/NetworkManager/NetworkManager.conf
 	fi
-	# point nameserver to 127.0.0.1
+	# Point nameserver to 127.0.0.1
 	sudo mv -f /etc/resolv.conf /etc/resolv.conf.orig
 	# Todo: add default router as nameserver
 	cat <<EOH | sudo tee /etc/resolv.conf
@@ -82,12 +84,14 @@ base_line() {
 nameserver 127.0.0.1
 EOH
 	sudo systemctl --now disable ModemManager 2>/dev/null
-	sudo systemctl --now disable wpa_supplicant
-	sudo systemctl --now disable nsd
-	sudo systemctl --now disable unbound
+	sudo systemctl --now disable wpa_supplicant 2>/dev/null
+	sudo systemctl --now disable nsd 2>/dev/null
+	sudo systemctl --now disable unbound 2>/dev/null
 	sudo systemctl enable NetworkManager
-	sudo systemctl restart NetworkManager
-	[ $? != 0 ] && echo "Error: Failed to start NetworkManager" && exit 1
+	if ! sudo systemctl restart NetworkManager; then
+		echo "Error: Failed to start NetworkManager"
+		exit 1
+	fi
 	sudo apt -y autoremove
 }
